+++ changelog.md
@@ -12,6 +12,7 @@
  * db_constraints.utils.meta.generic_compare -> db_constraints.utils.meta.opAAKey
  * db_constraints.utils.meta.HasMembersWithUDA -> db_constraints.utils.meta.hasMembersWithUDA
  * db_constraints.utils.meta.ConstraintStructs -> db_constraints.utils.meta.createConstraintStructs
+ * db_constraints.utils.meta.HasForeignKeys -> db_constraints.utils.meta.hasForeignKeys
 
 
 ### New additions